fix: resolve remaining cyclic-import alerts by extracting analysis_types
Extract `LatticeKind`, `LatticeValue`, and their helpers into a
new shared `compiler/analysis_types.py` module, breaking the
`compilation_unit` ↔ `core_analyses` ↔ `interprocedural` import
cycle at module level.

Changes:
- NEW: `compiler/analysis_types.py` — LatticeKind enum, LatticeValue
  dataclass, singletons (UNKNOWN/OVERDEFINED), and helpers
- `core_analyses.py`: imports Lattice types from analysis_types
- `interprocedural.py`: imports LatticeKind/Value from analysis_types;
  moves FunctionAnalysis to TYPE_CHECKING; defers analyse_function and
  \_expr_has_command to function-level imports
- `compilation_unit.py`: moves InterproceduralAnalysis,
  ProcLocalSummary, FunctionAnalysis to TYPE_CHECKING; defers
  analyse_interprocedural_ir and analyse_function to function level
- `tooling/explorer/tui.py`: defers all cli imports to function/
  method level to break the cli ↔ tui cycle
- Updated import paths in dataflow_graph, optimiser, taint, analyser,
  and minifier modules to use analysis_types

This resolves the remaining 11 cyclic-import alerts (9 from the
compiler pipeline + 2 from the explorer cli/tui cycle).
