+++ analyser/_analyser/_diag_ip.py
@@ -9,7 +9,8 @@ else:
     _Base = object
 
 from compiler.cfg import CFGFunction
-from compiler.core_analyses import FunctionAnalysis, LatticeKind
+from compiler.analysis_types import LatticeKind
+from compiler.core_analyses import FunctionAnalysis
 
 from ..semantic_model import Diagnostic, Range, Severity
